Reusable pytest fixtures.
