+++ tests/news_api/conftest.py
@@ -1,7 +1,7 @@
 import sys
 from pathlib import Path
 from pytest import fixture
-from tests.conftest import update_config # noqa
+from newsroom.tests.conftest import update_config # noqa
 
 root = (Path(__file__).parent / '..').resolve()
 sys.path.insert(0, str(root))
